Skip dry run on missing resources
When ArgoCD installs operator, the CRD are initally not there.
ArgoCD sync fails on object refering the not existing CRD.
In order to avoid this, the flag will tell ArgoCD to ignore the
dry run for the resource that have a none-existing Kind in the
cluster.

See https://argoproj.github.io/argo-cd/user-guide/sync-options/#skip-dry-run-for-new-custom-resources-types

Signed-off-by: Alexis de Talhouët <adetalhouet89@gmail.com>